[loong64] Fix the remaining implicit vector casts
CL 909436 fixed a few implicit vector casts, but they are not tested
with a proper Skia build (in fact it was mainly for fixing those errors
that happen to show up in Debian packaging e.g. wpewebkit), so many of
the implicit casts remain.

The changes here are tested with GCC 14, and confirmed to not affect
Clang builds. However, due to an issue with Clang headers [1],
-fno-lax-vector-conversions cannot be enabled for LoongArch Clang
builds yet, at least for Clang < 19.1.4 [2].

[1]: https://github.com/llvm/llvm-project/issues/110834
[2]: https://github.com/llvm/llvm-project/pull/114958

Change-Id: I3b4e9479cb6f9628b4cf796a0ac25098bc1836a2

[Kaiyang Wu: backport to webkit2gtk]

Link: https://skia-review.googlesource.com/c/skia/+/908137
Signed-off-by: Kaiyang Wu <origincode@aosc.io>